[Feat] loan 정렬 기준

+++ src/main/java/com/homeis/board/controller/BoardController.java
@@ -1,12 +1,13 @@
 package com.homeis.board.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,6 +17,7 @@ import com.homeis.board.dto.BoardPaginationResponse;
 import com.homeis.board.dto.Comment;
 import com.homeis.board.dto.Likes;
 import com.homeis.board.model.service.BoardService;
+import com.homeis.util.JWTUtil;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,6 +26,7 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/board")
 public class BoardController {
 	private final BoardService boardService;
+	private final JWTUtil jwtUtil;
 	
 	@GetMapping("/list")
 	public ResponseEntity<BoardPaginationResponse> list(
@@ -37,15 +40,26 @@ public class BoardController {
 	}
 	
 	@GetMapping("/detail/{id}")
-	public ResponseEntity<Board> detail(@PathVariable("id") int id) {
-		Board board = boardService.findById(id);
+	public ResponseEntity<?> detail(@PathVariable("id") int id, @RequestHeader(value = "Authorization", required = false) String tokenHeader) {
+		String tokenId = null;
 		
-		if(board == null) ResponseEntity.status(404).body("요청하신 글을 찾을 수 없습니다.");
+		if(tokenHeader != null && tokenHeader.startsWith("Bearer ")) {
+			tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		}
+		
+		Board board = boardService.findById(id, tokenId);
+		
+		if(board == null) return ResponseEntity.status(404).body("요청하신 글을 찾을 수 없습니다.");
+		
 		return ResponseEntity.ok(board);
 	}
 	
 	@PostMapping("/insert")
-	public ResponseEntity<?> regist(@RequestBody Board board) {
+	public ResponseEntity<?> regist(@RequestBody Board board, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertBoard(board);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -54,25 +68,37 @@ public class BoardController {
 	}
 	
 	@PutMapping("/update")
-	public ResponseEntity<?> update(@RequestBody Board board) {
-		int result = boardService.updateBoard(board);
+	public ResponseEntity<?> update(@RequestBody Board board, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
 		
-		if(result == 0) return ResponseEntity.notFound().build();
+		int result = boardService.updateBoard(board);
+
+		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
 		
 		return ResponseEntity.ok().build();
 	}
 	
-	@DeleteMapping("/delete")
-	public ResponseEntity<?> delete(@RequestBody Board board) {
+	@PutMapping("/delete")
+	public ResponseEntity<?> delete(@RequestBody Board board,@RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteBoard(board);
 		
-		if(result == 0) return ResponseEntity.notFound().build();
+		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
 		
 		return ResponseEntity.ok().build();
 	}
 	
 	@PostMapping("/like")
-	public ResponseEntity<?> likeRegist(@RequestBody Likes like) {
+	public ResponseEntity<?> likeRegist(@RequestBody Likes like, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!like.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertLike(like);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -80,8 +106,12 @@ public class BoardController {
 		return ResponseEntity.ok().build();
 	}
 	
-	@DeleteMapping("/like")
-	public ResponseEntity<?> likeDelete(@RequestBody Likes like) {
+	@PutMapping("/like")
+	public ResponseEntity<?> likeDelete(@RequestBody Likes like, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!like.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteLike(like);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -90,7 +120,11 @@ public class BoardController {
 	}
 	
 	@PostMapping("/insert-comment")
-	public ResponseEntity<?> commentRegist(@RequestBody Comment comment) {
+	public ResponseEntity<?> commentRegist(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertComment(comment);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -99,19 +133,27 @@ public class BoardController {
 	}
 	 
 	@PutMapping("/update-comment")
-	public ResponseEntity<?> commentupdate(@RequestBody Comment comment) {
+	public ResponseEntity<?> commentupdate(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.updateComment(comment);
 		
-		if(result == 0) return ResponseEntity.notFound().build();
+		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
 		
 		return ResponseEntity.ok().build();
 	}
 	
-	@DeleteMapping("/delete-comment")
-	public ResponseEntity<?> commentDelete(@RequestBody Comment comment) {
+	@PutMapping("/delete-comment")
+	public ResponseEntity<?> commentDelete(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteComment(comment);
 		
-		if(result == 0) return ResponseEntity.notFound().build();
+		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
 		
 		return ResponseEntity.ok().build();
 	}
